Keep server composition modules within existing line budgets
