+++ services/server/src/features/grants/policy.ts
@@ -1,4 +1,5 @@
 import { randomUUID } from "node:crypto";
+import { isDeepStrictEqual } from "node:util";
 import {
   APPLICATION_SETUP_OPERATIONS,
   applicationOperationSelectionIsAtomic,
@@ -12,6 +13,7 @@ import type {
   ApplicationProvisions,
   CollectionContractDescriptor,
   ContractRequirement,
+  ContractSetupChoice,
   GrantEncryption,
   GrantScope,
   TypePackProvision
@@ -135,6 +137,78 @@ export function requiredContractsForRequirements(
     `${contract.id}@${contract.version}`,
     contract
   ])).values()];
+}
+
+export function verifyContractSetupAcknowledgement(
+  requested: ContractSetupChoice[],
+  acknowledged: ContractSetupChoice[] | undefined,
+  contracts: CollectionContractDescriptor[]
+): void {
+  if (requested.length === 0) return;
+  if (!acknowledged || !isDeepStrictEqual(acknowledged, requested)) {
+    throw new RequestValidationError(
+      "The collection authority did not acknowledge the exact contract setup that was approved."
+    );
+  }
+  for (const setup of requested) {
+    const contract = contracts.find((candidate) =>
+      candidate.id === setup.contract.id
+      && candidate.version === setup.contract.version
+    );
+    if (!contract) {
+      throw new RequestValidationError(
+        `Contract setup did not provide ${setup.contract.id} ${setup.contract.version}.`
+      );
+    }
+    if (setup.mode === "starter") continue;
+    const implementation = contract.implementations.find((candidate) =>
+      candidate.type_name === setup.type_name
+      && isDeepStrictEqual(candidate.fields, setup.fields)
+      && isDeepStrictEqual(candidate.binding, setup.binding)
+    );
+    if (!implementation) {
+      throw new RequestValidationError(
+        `Contract setup did not apply the approved mapping to type '${setup.type_name}'.`
+      );
+    }
+  }
+}
+
+export function validateContractSetupChoices(
+  setups: ContractSetupChoice[],
+  required: ContractRequirement[],
+  available: CollectionContractDescriptor[]
+): void {
+  const keys = new Set(setups.map(
+    (setup) => `${setup.contract.id}@${setup.contract.version}#${setup.contract.digest}`
+  ));
+  if (
+    keys.size !== setups.length
+    || setups.some((setup) => !required.some((contract) =>
+      contract.id === setup.contract.id
+        && contract.version === setup.contract.version
+        && contract.digest === setup.contract.digest
+    ))
+  ) {
+    throw new RequestValidationError(
+      "Contract setup may configure each contract required by this application only once."
+    );
+  }
+  if (setups.length === 0) return;
+  const missing = required.filter((contract) => !available.some((candidate) =>
+    candidate.id === contract.id
+      && candidate.version === contract.version
+      && candidate.digest === contract.digest
+  ));
+  if (
+    keys.size !== missing.length
+    || missing.some((contract) =>
+      !keys.has(`${contract.id}@${contract.version}#${contract.digest}`))
+  ) {
+    throw new RequestValidationError(
+      "Choose starter or existing-type setup for each missing contract only."
+    );
+  }
 }
 
 export function allowedTypesForRequirements(
